+++ homepage/homepage/app/(docs)/docs/[framework]/page.tsx
@@ -1,6 +1,5 @@
 import { Framework, frameworks } from "@/content/framework";
-import { DocPage, getDocMetadata } from "@/lib/docMdxContent";
-import { generateOGMetadata } from "@/lib/metadata";
+import { DocPage, generateOGMetadata, getDocMetadata } from "@/lib/docMdxContent";
 
 export async function generateMetadata({
   params,
