+++ homepage/homepage/lib/docMdxContent.tsx
@@ -177,3 +177,37 @@ function filterTocItemsForFramework(
     })
     .filter(Boolean) as Toc;
 }
+
+export function generateOGMetadata(
+  framework: string,
+  slug: string[],
+  docMeta: { title: string; description: string; image?: string; topic?: string; subtopic?: string }
+) {
+  const { title, description, image, topic, subtopic } = docMeta;
+
+  return {
+    title,
+    description,
+    openGraph: {
+      title,
+      description,
+      type: "article",
+      url: `https://jazz.tools/docs/${[framework, ...slug].join("/")}`,
+      images: [
+        {
+          url: `/opengraph-image?title=${encodeURIComponent(title)}&framework=${encodeURIComponent(framework)}`,
+          width: 1200,
+          height: 630,
+          alt: title,
+        },
+      ],
+    },
+    twitter: {
+      card: "summary_large_image",
+      title,
+      description,
+      images: [image],
+    },
+  };
+}
+
